refactor(sample): move CODE_SAMPLES to assets/samples/ files
Replace 700+ lines of hardcoded Kotlin string literals in SampleData.kt
with individual asset files under sample/src/main/assets/samples/.

Each file uses a real extension (.py, .kt, .rs, etc.) so editors and
IDEs apply proper syntax highlighting when viewing or editing samples.
New samples can be added by dropping a file in the folder — no Kotlin
changes needed.

Changes:
- Add 17 asset files with numeric prefix for ordering (01_fibonacci.py
  through 17_plaintext.txt)
- Add loadCodeSamples(Context) and extensionToLanguage() helpers
- Add displayLabel field to CodeSample to separate the human-readable
  filename from the hljs language identifier
- Fix latent bug: 'kotlin (large — WeatherApp)' was passed to hljs as
  the language ID, causing silent no-op highlighting; now uses 'kotlin'
- Update SampleScreen and PerfScreen to load samples at composable scope
  via remember(context) { loadCodeSamples(context) }
