+++ sample/src/main/assets/samples/05_query.sql
@@ -1,0 +1,12 @@
+SELECT
+    u.id,
+    u.name,
+    COUNT(o.id) AS order_count,
+    SUM(o.total) AS revenue
+FROM users u
+LEFT JOIN orders o ON o.user_id = u.id
+WHERE u.created_at >= '2024-01-01'
+GROUP BY u.id, u.name
+HAVING order_count > 0
+ORDER BY revenue DESC
+LIMIT 10;
