+++ sample/src/main/assets/samples/16_styles.css
@@ -1,0 +1,46 @@
+:root {
+    --color-primary:  #6200ea;
+    --color-surface:  #ffffff;
+    --color-on-surface: #1c1b1f;
+    --radius-md: 8px;
+    --transition: 200ms cubic-bezier(0.4, 0, 0.2, 1);
+}
+
+@media (prefers-color-scheme: dark) {
+    :root {
+        --color-primary:    #bb86fc;
+        --color-surface:    #121212;
+        --color-on-surface: #e6e1e5;
+    }
+}
+
+.card {
+    background: var(--color-surface);
+    border-radius: var(--radius-md);
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.12);
+    color: var(--color-on-surface);
+    padding: 1.5rem;
+    transition: transform var(--transition), box-shadow var(--transition);
+}
+
+.card:hover {
+    transform: translateY(-2px);
+    box-shadow: 0 6px 20px rgba(0, 0, 0, 0.18);
+}
+
+@keyframes fadeIn {
+    from { opacity: 0; transform: translateY(6px); }
+    to   { opacity: 1; transform: translateY(0);   }
+}
+
+.card--animated {
+    animation: fadeIn 0.35s var(--transition) both;
+}
+
+.card__title {
+    color: var(--color-primary);
+    font-size: clamp(1rem, 2.5vw, 1.5rem);
+    font-weight: 600;
+    letter-spacing: 0.01em;
+    margin: 0 0 0.5rem;
+}
