+++ sample/src/main/assets/samples/17_plaintext.txt
@@ -1,0 +1,2 @@
+This is plaintext.
+Nothing to highlight.
